bayes opt code
